add correct user check for /api/set/delete and /api/set/update

+++ website/backend/src/routes/flashcardset.js
@@ -21,9 +21,6 @@ router.route("/").get(auth, (req, res) => {
       }  
     })
     .then((user) => {
-      if (!user) {
-        return res.status(400).json({ error: "User not found." });
-      }
       let newSets = [];
 
       user.sets.forEach((set, idx) => {
@@ -87,21 +84,27 @@ router.route("/delete/:id").delete(auth, (req, res) => {
       if (!set) {
         return res.status(400).json({error: `Flashcard set with id ${id} does not exist`});
       }
-
-      // If set exists, delete all flashcards that belong to set AND deletes entry in user that had the flashcard set.
-      Promise.all([...set.flashcards.map((flashcardId) => Flashcard.findByIdAndDelete(flashcardId).exec()),
-        User.findOneAndUpdate({_id: mongoose.Types.ObjectId(user.id)}, {
-          $pull: {sets: mongoose.Types.ObjectId(id)}
-          }, 
-            (err) => {
-              if (err) {
-                return res.status(500).json({error: err});
-              }
+      User.findById(user.id)
+        .then((user) => {
+            if (set.user !== user.username) {
+                return res.status(403).json({error: `User with id ${user.id} is not the creator of this set.`});
             }
-          )])
-        .then((_) => FlashcardSet.findByIdAndDelete(req.params.id))
-        .then((_) => res.json({msg: "Flashcard Set deleted"}))
-        .catch(err => res.status(400).json({error: err}));
+            // If set exists, delete all flashcards that belong to set AND deletes entry in user that had the flashcard set.
+            Promise.all([...set.flashcards.map((flashcardId) => Flashcard.findByIdAndDelete(flashcardId).exec()),
+                User.findOneAndUpdate({_id: mongoose.Types.ObjectId(user.id)}, {
+                $pull: {sets: mongoose.Types.ObjectId(id)}
+                }, 
+                    (err) => {
+                        if (err) {
+                            return res.status(500).json({error: err});
+                        }
+                    }
+                )])
+                .then((_) => FlashcardSet.findByIdAndDelete(req.params.id))
+                .then((_) => res.json({msg: "Flashcard Set deleted"}))
+                .catch(err => res.status(400).json({error: err}));
+        });
+
     })
     .catch((err) => res.status(500).json({error: err}));
 
@@ -119,40 +122,38 @@ router.route("/create").post(auth, (req, res) => {
   }
   User.findById(user.id)
     .then((user) => {
-      if (!user) {
-        return res.status(400).json({ error: "User not found." });
-      }
-      else {
-        let flashcardIds = [];
-        let flashcardDocs = [];
-        if (flashcards) {
-          flashcardDocs = flashcards.map((flashcard) => new Flashcard(flashcard));
-          flashcardIds = flashcardDocs.map((flashcard) => flashcard._id);
-        }
+      let flashcardIds = [];
+      let flashcardDocs = [];
 
-        Promise.all(flashcardDocs.map((flashcard) => flashcard.save()))
-          .then((_) => {
-            const newSet = new FlashcardSet({
-              name,
-              flashcards: flashcardIds,
-              user: user.username
-            });
-    
-            user.sets.push(mongoose.Types.ObjectId(newSet._id));
-    
-            Promise.all([user.save(), newSet.save()])
-              .then((_) => res.json(newSet))
-              .catch((err) => res.status(500).json({ error: err }));
-          });
+      if (flashcards) {
+        flashcardDocs = flashcards.map((flashcard) => new Flashcard(flashcard));
+        flashcardIds = flashcardDocs.map((flashcard) => flashcard._id);
       }
-    })
+
+      Promise.all(flashcardDocs.map((flashcard) => flashcard.save()))
+        .then((_) => {
+          const newSet = new FlashcardSet({
+            name,
+            flashcards: flashcardIds,
+            user: user.username
+          });
+  
+          user.sets.push(mongoose.Types.ObjectId(newSet._id));
+  
+          Promise.all([user.save(), newSet.save()])
+            .then((_) => res.json(newSet))
+            .catch((err) => res.status(500).json({ error: err }));
+        });
+    }
+    )
     .catch((err) => res.status(500).json({ error: err }));
 
   
 })
 
 // Handle updating a flashcard set
-router.route("/update/:id").patch((req, res) => {
+router.route("/update/:id").patch(auth, (req, res) => {
+  const { user } = req;
   const { name } = req.body;
   const { id } = req.params;
   
@@ -161,17 +162,25 @@ router.route("/update/:id").patch((req, res) => {
   }
 
   FlashcardSet.findById(id)
-  .then((card) => {
-    if (!card) {
-      return res.status(400).json({ error: `Flashcard set with id ${id} does not exist` });
+  .then((set) => {
+    if (!set) {
+      return res.status(404).json({ error: `Flashcard set with id ${id} does not exist` });
     }
 
-    card.name = name;
-    card.save()
-    .then((_) => {
-      res.json(card);
-    })
-    .catch((err) => res.status(500).json({error: err}));
+    User.findById(user.id)
+      .then((user) => {
+        if (user.username !== set.user) {
+          return res.status(403).json({error: `User with id ${user} is not the creator of this set.`});
+        }
+
+        set.name = name;
+        set.save()
+        .then((_) => {
+          res.json(set);
+        })
+        .catch((err) => res.status(500).json({error: err}));
+      });
+    
   })
   .catch((err) => res.status(500).json({error: err}));
 });
